Move and rename some things
